Refactored API handlers into separate module and updated GuideAPI

+++ models/url.go
@@ -2,5 +2,5 @@ package models
 
 type URL struct {
     Address string
-    Priority int
+    Priority uint8
 }
